+++ test/invalid_spec_test.go
@@ -5,7 +5,7 @@ package tests
 import (
 	"context"
 
-	//	"github.com/pulumi/pulumi-kubernetes-operator/pkg/apis/pulumi/shared"
+	"github.com/pulumi/pulumi-kubernetes-operator/pkg/apis/pulumi/shared"
 	pulumiv1 "github.com/pulumi/pulumi-kubernetes-operator/pkg/apis/pulumi/v1"
 
 	. "github.com/onsi/ginkgo"
@@ -15,28 +15,19 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-var _ = Describe("Stacks that should stall", func() {
-
-	var (
-		stack pulumiv1.Stack
-	)
-
-	AfterEach(func() {
-		if stack.Name != "" { // assume that if it's been named, it was created in the cluster
-			Expect(k8sClient.Delete(context.TODO(), &stack)).To(Succeed())
-		}
-	})
-
-	When("there is no source specified", func() {
+func checkInvalidSpecStalls(when string, setup func(stack *pulumiv1.Stack)) {
+	When(when, func() {
+		var stack pulumiv1.Stack
 
 		BeforeEach(func() {
 			stack = pulumiv1.Stack{}
 			stack.Name = randString()
 			stack.Namespace = "default"
+			setup(&stack)
 			Expect(k8sClient.Create(context.TODO(), &stack)).To(Succeed())
 		})
 
-		It("should mark a stack as stalled", func() {
+		It("should mark the stack as stalled", func() {
 			// wait until the controller has seen the stack object and completed processing it
 			var s pulumiv1.Stack
 			Eventually(func() bool {
@@ -49,12 +40,43 @@ var _ = Describe("Stacks that should stall", func() {
 				}
 				return s.Status.ObservedGeneration == s.Generation
 			}, "20s", "1s").Should(BeTrue())
-
-			Expect(apimeta.IsStatusConditionTrue(s.Status.Conditions, pulumiv1.ReadyCondition)).To(BeFalse())
-			Expect(apimeta.FindStatusCondition(s.Status.Conditions, pulumiv1.ReconcilingCondition)).To(BeNil())
+			Expect(s.Status.LastUpdate).ToNot(BeNil(), ".status.lastUpdate is recorded")
+			Expect(s.Status.LastUpdate.State).To(Equal(shared.FailedStackStateMessage))
 			stalledCondition := apimeta.FindStatusCondition(s.Status.Conditions, pulumiv1.StalledCondition)
-			Expect(stalledCondition).ToNot(BeNil())
+			Expect(stalledCondition).ToNot(BeNil(), "stalled condition is present")
 			Expect(stalledCondition.Reason).To(Equal(pulumiv1.StalledSpecInvalidReason))
+			// not ready, and not in progress
+			Expect(apimeta.IsStatusConditionTrue(s.Status.Conditions, pulumiv1.ReadyCondition)).To(BeFalse(), "ready condition is false")
+			Expect(apimeta.FindStatusCondition(s.Status.Conditions, pulumiv1.ReconcilingCondition)).To(BeNil(), "reconciling condition is absent")
 		})
+
+		AfterEach(func() {
+			if stack.Name != "" { // assume that if it's been named, it was created in the cluster
+				Expect(k8sClient.Delete(context.TODO(), &stack)).To(Succeed())
+			}
+		})
+	})
+}
+
+var _ = Describe("Stacks that should stall because of an invalid spec", func() {
+
+	checkInvalidSpecStalls("there is no source specified", func(*pulumiv1.Stack) {})
+
+	checkInvalidSpecStalls("a git source with no branch and no commit hash is given", func(s *pulumiv1.Stack) {
+		s.Spec.GitSource = &shared.GitSource{
+			ProjectRepo: "https://github.com/pulumi/pulumi-kubernetes-operator",
+		}
+	})
+
+	checkInvalidSpecStalls("more than one source is given", func(s *pulumiv1.Stack) {
+		s.Spec.GitSource = &shared.GitSource{
+			ProjectRepo: "https://github.com/pulumi/pulumi-kubernetes-operator",
+			Branch:      "default",
+		}
+		s.Spec.FluxSource = &shared.FluxSource{
+			SourceRef: shared.FluxSourceReference{
+				Name: "foo",
+			},
+		}
 	})
 })